deps/v8/test/inspector/debugger/class-private-methods-empty-inner-expected.txt

Test empty inner classes with private instance methods in the outer class

Running test: testScopesPaused
undefined
